+++ hashfunction/constants.go
@@ -1,0 +1,13 @@
+package main
+
+
+const (
+    c1           uint64 = 0xcc9e2d51
+    c2           uint64 = 0x1b873593
+    Prime1       uint64 = 0x85ebca6b
+    Prime2       uint64 = 0xc2b2ae35
+    Prime3       uint64 = 0xe6546b64
+    base62              = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
+    base62Length        = 62
+    seed         uint64 = 0x9747b28c
+)
