--- main.go
@@ -1,8 +1,0 @@
-package main
-
-import HashFunction "github.com/gurnoorsinghbhinder/brevo-assignments/hashfunction"
-
-func main(){
-	HashFunction.HashFunction()
-	
-}
